delete duplicate classpath entries
When using ensime with multiple submodules with overlapping
dependencies, I was finding that the exact same jars were being added
multiple times to the classpath. Java doesn't consider this an issue,
but the classpath was long enough due to this issue to exceed the
maximum size allowed for bash on my machine (argument list too long exception).