Grouped scala import insertion w/ common package.
Adds functionality to import insertions to group a current import in the
buffer and the desired import together if they share the same base
package.

i.e.

If there is already a statement of
`import com.example.MyCoolClass`

and a user is trying to import `com.example.MyOtherCoolClass`,

the result will be

`import com.example.{ MyCoolClass, MyOtherCoolClass }`

whereas prior to this commit, the import statement would have been added
after on a separate line.