Shrink search entry height when window space is limited
After the search window is created it is split into a results window and
a search entry window. The entry window is generally 4 lines high. If
window space is limited then use a smaller entry area.

This will still blow up (just like the original implementation) if there
is not enough room to create the entry area. This implementation just
makes that a little less likely to happen.

The only place this is likely to come into play is the for unit tests in
ensime-test.el. Those run in emacs --batch mode where the frame-height
is set to some very small value (e.g. 9 or 10 lines).